RELATED: FET-203 Remove circular dependencies part 1
