+++ package.json
@@ -67,6 +67,7 @@
     "isomorphic-fetch": "2.2.1",
     "jest": "24.7.1",
     "jest-junit": "3.0.0",
+    "madge": "^3.4.4",
     "moment": "2.23.0",
     "node-sass": "4.7.2",
     "node-sass-magic-importer": "5.1.0",
@@ -150,6 +151,7 @@
     "tslint-ci": "mkdir -p ./ci/results && yarn run-tslint -t checkstyle -o ./ci/results/tslint-results.xml $npm_package_config_tslint",
     "prettier-check": "prettier --check '{src,stories,styles,__mocks__,examples/server/src,examples/src,examples/test}/**/*.{ts,js,tsx,jsx,json,scss,md,yaml,html}'",
     "prettier-write": "prettier --write '{src,stories,styles,__mocks__,examples/server/src,examples/src,examples/test}/**/*.{ts,js,tsx,jsx,json,scss,md,yaml,html}'",
+    "check-circular-deps": "madge --circular --extensions ts ./src",
     "validate": "tsc --noEmit && yarn tslint && eslint $npm_package_config_eslint && yarn prettier-check",
     "validate-ci": "tsc --noEmit && yarn tslint-ci  && eslint $npm_package_config_eslint -f checkstyle -o ci/results/eslint-results.xml && yarn prettier-check",
     "storybook": "yarn build && start-storybook -p 9001 -c .storybook",
