+++ src/components/visualizations/table/constants/layout.ts
@@ -1,0 +1,7 @@
+// (C) 2019 GoodData Corporation
+export const DEFAULT_HEADER_HEIGHT: number = 26;
+export const DEFAULT_ROW_HEIGHT: number = 30;
+export const DEFAULT_FOOTER_ROW_HEIGHT: number = 30;
+
+export const TOTALS_ADD_ROW_HEIGHT: number = 50;
+export const TOTALS_TYPES_DROPDOWN_WIDTH: number = 150;
